+++ public/index.php
@@ -1,0 +1,5 @@
+<?php
+
+include_once __DIR__ . "/../vendor/autoload.php";
+$app = (new App\Composer\Core\Application)->getInstance();
+$app->run();
